LayoutTests/inspector/console/console-assert.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/console-test.js"></script>
<script>

function b()
{
    console.assert(false, 1);
    console.assert(false, "a", "b");
}

function a()
{
    b();
}

function test()
{
    var callCount = 0;
    function callback()
    {
        if (++callCount === 2) {
            InspectorTest.expandConsoleMessages();
            InspectorTest.dumpConsoleMessages();
            InspectorTest.completeTest();
        }
    }
    InspectorTest.evaluateInPage("setTimeout(a, 0)");
    InspectorTest.addConsoleSniffer(callback, true);
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that console.assert() will dump a message and stack trace with source URLs and line numbers.
</p>

</body>
</html>
