LayoutTests/inspector/console/console-assert-expected.txt

CONSOLE MESSAGE: line 9: 1
CONSOLE MESSAGE: line 10: a
Tests that console.assert() will dump a message and stack trace with source URLs and line numbers.

Assertion failed: 1 console-assert.html:9b console-assert.html:9a console-assert.html:15
Assertion failed: a b console-assert.html:10b console-assert.html:10a console-assert.html:15
